BPM-56 Refactored utils into cookie class
